Update #4
This is the biggest milestone of an update in the development of this project. It adds countdowns, better ball spawning, loading between juggle and menu scenes, allowing data on number of starting balls to persist from menu to juggle scene, and (largest addition of all) saving the highest score to an external file and loading that score between plays. This update also includes some bug fixes with systems already in place, such as the OutOfBounds trigger and end game.